Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- net_(delete/update)_notify methods modified,to update dicts on each nodes.

- ipam_(delete/update)_notify modified too.

- new parameter added (notify) now, "IndexAllocator.delete_all"
  is done only from "*_req" methods.

- unittests updated to match and test new behavior

Closes-Bug: #1332645
Closes-Bug: #1365824

+++ src/config/common/vnc_rdbms.py
@@ -1520,10 +1520,12 @@ class VncRDBMSClient(object):
                 max_alloc=self._MAX_SUBNET_ADDR_ALLOC/alloc_unit)
     # end create_subnet_allocator
 
-    def delete_subnet_allocator(self, subnet):
-        self._subnet_allocators.pop(subnet, None)
-        RDBMSIndexAllocator.delete_all(self.db,
-                                  self._subnet_path+'/'+subnet+'/')
+    def delete_subnet_allocator(self, subnet, notify=True):
+        if subnet in self._subnet_allocators:
+            self._subnet_allocators.pop(subnet, None)
+        if not notify:
+            RDBMSIndexAllocator.delete_all(self.db,
+                    self._subnet_path+'/'+subnet+'/')
     # end delete_subnet_allocator
 
     def _get_subnet_allocator(self, subnet):
